requirements.txt
requests~=2.28.2
prefect~=2.8.6
beautifulsoup4~=4.11.2
python-dotenv~=1.0.0